fix: check for ownership of franchise before deleting

+++ src/routes/franchiseRouter.js
@@ -98,6 +98,11 @@ franchiseRouter.delete(
   '/:franchiseId',
   asyncHandler(async (req, res) => {
     const franchiseId = Number(req.params.franchiseId);
+    const franchise = await DB.getFranchise({ id: franchiseId });
+    if (!franchise || (!req.user.isRole(Role.Admin) && !franchise.admins.some((admin) => admin.id === req.user.id))) {
+        throw new StatusCodeError('unable to delete a store', 403);
+    }
+
     await DB.deleteFranchise(franchiseId);
     res.json({ message: 'franchise deleted' });
   })
